backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-docx==1.1.0
pdfplumber==0.10.3
spacy==3.7.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
openai==1.10.0
anthropic==0.18.1
pydantic==2.5.3
aiofiles==23.2.1
requests==2.31.0
